Added corrections from @matz-e and @pramodk
